added test around remove users and groups

+++ manifests/remove.pp
@@ -16,26 +16,29 @@ class local_users::remove (
     $groups_to_add_list = keys( $local_users::groups_to_add )
   }
 
-  $users_to_remove = $local_users::users_to_remove - $users_to_add_list
-  $groups_to_remove = $local_users::groups_to_remove - $groups_to_add_list
-
-  $users_to_remove.each | $user | {
-    exec { "killing ${user}":
-      command => "pkill -9 -u ${user}",
-      onlyif  => "grep '^${user}' /etc/passwd && ps -u ${user}",
-      path    => [ '/bin', '/sbin', '/usr/bin', '/usr/sbin' ],
-    }
-    -> user { $user:
-      ensure     => absent,
-      forcelocal => $local_users::forcelocal,
-      managehome => $local_users::managehome,
+  if !empty( $local_users::users_to_remove ) {
+    $users_to_remove = $local_users::users_to_remove - $users_to_add_list
+    $users_to_remove.each | $user | {
+      exec { "killing ${user}":
+        command => "pkill -9 -u ${user}",
+        onlyif  => "grep '^${user}' /etc/passwd && ps -u ${user}",
+        path    => [ '/bin', '/sbin', '/usr/bin', '/usr/sbin' ],
+      }
+      -> user { $user:
+        ensure     => absent,
+        forcelocal => $local_users::forcelocal,
+        managehome => $local_users::managehome,
+      }
     }
   }
 
-  $groups_to_remove.each | $group | {
-    group { $group:
-      ensure     => absent,
-      forcelocal => $local_users::forcelocal,
+  if !empty( $local_users::groups_to_remove ) {
+    $groups_to_remove = $local_users::groups_to_remove - $groups_to_add_list
+    $groups_to_remove.each | $group | {
+      group { $group:
+        ensure     => absent,
+        forcelocal => $local_users::forcelocal,
+      }
     }
   }
 }
